refactor: move the inbox and nudge cluster into inbox.go
A pure move, no behaviour change: relay.go was 1141 lines against the
project's 800 cap after regrowing since its last split.
